Added in QDebug. Changed updateGraphic to addGraphic in GCSGraphicsLayer. Changed insertPath to lineTo. Added two files to the debug folder: libeay32.dll and ssleay32.dll.

+++ code/gcs/src/protonet/networking.h
@@ -98,7 +98,7 @@ signals:
     void targeting(float,float,float);
 
     //targeting
-    void manTargeting(float,float,float);
+    void manTargeting(double,double,double);
 
     //from vech
     void vTarget(float,float);
